moved PeriodeEncodageOuverte to commondomain
