+++ ddd/logic/encodage_des_notes/common_domain/service/periode_encodage_ouverte.py
